+++ website_formio/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 16.0.1.1
+
+Enable a portal and/or public Form to _redirect to a configured website page_, upon a _Save Draft is Done_.\
+This can be set in new fields in the Form Builder:
+- Portal Save-Draft Done Page
+- Public Save-Draft Done Page
+
 ## 16.0.1.0
 
 Initial release.